Back-end/requirement.txt
annotated-types==0.6.0
asgiref==3.7.2
black==24.1.1
click==8.1.7
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
